LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-gainnode-interface/no-dezippering-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.

Harness Error (FAIL), message = Unhandled rejection: The object is in an invalid state.

PASS # AUDIT TASK RUNNER STARTED. 
FAIL Executing "test0" promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: ConstantSourceNode"
PASS Executing "test2" 
PASS Executing "test3" 
PASS Audit report 
PASS > [test0] Dezippering of GainNode removed 
PASS > [test2] Compare value setter and setValueAtTime 
FAIL X .value setter output expected to be equal to the array [0,0.16793830692768097,0.3311063051223755,0.48486924171447754,0.6248595118522644,0.7471005916595459,0.8481203317642212,0.925049364566803,0.9757022261619568,0.9986401796340942,0.9932120442390442,0.9595714211463928,0.898674488067627,0.812250554561615,0.7027547955513,0.5732971429824829...] but differs in 640 places:
	Index	Actual			Expected
	[384]	9.2387950420379639e-1	2.7716386318206787e-1
	[385]	8.4649103879928589e-1	2.5394731760025024e-1
	[386]	7.4505776166915894e-1	2.2351734340190887e-1
	[387]	6.2246131896972656e-1	1.8673840165138245e-1
	...and 636 more errors. assert_true: expected true got false
FAIL X .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test2] 2 out of 2 assertions were failed. assert_true: expected true got false
PASS > [test3] Automation effects 
FAIL X With modulation: .value setter output expected to be equal to the array [0,0.23077477514743805,0.5608845949172974,0.9275365471839905,1.2492955923080444,1.4497531652450562,1.480018973350525,1.3333395719528198,1.047479271888733,0.6941256523132324,0.3584260046482086,0.11469673365354538,0.00547564635053277,0.03009113110601902,0.14609801769256592,0.2831018567085266...] but differs in 640 places:
	Index	Actual			Expected
	[384]	1.2774327993392944e+0	6.3071721792221069e-1
	[385]	8.5428196191787720e-1	2.6173824071884155e-1
	[386]	4.7265496850013733e-1	-4.8885431140661240e-2
	[387]	1.9463413953781128e-1	-2.4108877778053284e-1
	...and 636 more errors. assert_true: expected true got false
FAIL X With modulation: .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test3] 2 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 2 out of 3 tasks were failed. assert_true: expected true got false
